Code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "studybros"
version = "0.1.0"
description = "AI-powered study guide generator - transforms learning materials into flashcards, quizzes, and practice tests using any LLM via MCP"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "StudyBros Team"}
]
keywords = ["education", "ai", "flashcards", "quiz", "study", "learning", "mcp", "spaced-repetition"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-pptx>=0.6.21",
    "pypdf>=3.0.0",
    "openai>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
study-guide = "study_guide.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["study_guide*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
